requirements.txt
huggingface-hub==0.0.13
regex==2022.4.17
Pillow~=10.1.0
transformers~=4.36.0.dev0
datasets~=2.12.0
torch==1.10.0
flask~=2.2.2